change fabric mem alloc
