+++ mooncake-store/src/CMakeLists.txt
@@ -115,21 +115,6 @@ if(URING_LIB AND URING_INCLUDE)
     target_include_directories(mooncake_store PRIVATE ${URING_INCLUDE})
 endif()
 
-if (USE_ASCEND_DIRECT OR USE_UBSHMEM)
-    set(ACL_RUNTIME_HEADER_PATH "${ASCEND_INCLUDE_DIR}/acl/acl_rt.h")
-    if(EXISTS "${ACL_RUNTIME_HEADER_PATH}")
-        file(READ "${ACL_RUNTIME_HEADER_PATH}" ACL_RUNTIME_CONTENT)
-        if("${ACL_RUNTIME_CONTENT}" MATCHES "ACL_MEM_P2P_HUGE1G")
-            message(STATUS "ACL_MEM_P2P_HUGE1G exist.")
-            add_compile_definitions(ASCEND_SUPPORT_FABRIC_MEM)
-        else()
-            message(STATUS "ACL_MEM_P2P_HUGE1G is not exist.")
-        endif()
-    else()
-        message(WARNING "Acl runtime header file is not exist")
-    endif()
-endif()
-
 if (BUILD_SHARED_LIBS)
   install(TARGETS mooncake_store DESTINATION lib)
 endif()
