+++ mooncake-store/src/utils.cpp
@@ -13,10 +13,10 @@
 #include <csignal>
 #include <cstring>
 #include <sys/mman.h>
-#if defined(USE_ASCEND_DIRECT) || defined(USE_UBSHMEM)
-#include "acl/acl.h"
 #include "config.h"
 #include "common.h"
+#if defined(USE_ASCEND_DIRECT) || defined(USE_UBSHMEM)
+#include "ascend_allocator.h"
 #endif
 
 #include <ylt/coro_http/coro_http_client.hpp>
@@ -75,48 +75,6 @@ AutoPortBinder::~AutoPortBinder() {
     }
 }
 
-#if (defined(USE_ASCEND_DIRECT) || defined(USE_UBSHMEM)) && \
-    defined(ASCEND_SUPPORT_FABRIC_MEM)
-int allocate_physical_memory(size_t total_size, aclrtDrvMemHandle &handle) {
-    int32_t user_dev_id;
-    auto ret = aclrtGetDevice(&user_dev_id);
-    if (ret != ACL_ERROR_NONE) {
-        LOG(ERROR) << "Failed to get device: " << ret;
-        return -1;
-    }
-    int32_t physical_dev_id;
-    ret = aclrtGetPhyDevIdByLogicDevId(user_dev_id, &physical_dev_id);
-    if (ret != ACL_ERROR_NONE) {
-        LOG(ERROR) << "Failed to get physical dev id: " << ret;
-        return -1;
-    }
-    aclrtPhysicalMemProp prop = {};
-    prop.handleType = ACL_MEM_HANDLE_TYPE_NONE;
-    prop.allocationType = ACL_MEM_ALLOCATION_TYPE_PINNED;
-    prop.memAttr = ACL_MEM_P2P_HUGE1G;
-    prop.location.type = ACL_MEM_LOCATION_TYPE_HOST_NUMA;
-    // Only 0 2 4 6 is available for fabric mem, map 4 device to one numa.
-    const int32_t kDevicesPerChip = 4;
-    const int32_t kNumaNodeStep = 2;
-    prop.location.id = (physical_dev_id / kDevicesPerChip) * kNumaNodeStep;
-    prop.reserve = 0;
-    LOG(INFO) << "Malloc host memory for numa:" << prop.location.id;
-    ret = aclrtMallocPhysical(&handle, total_size, &prop, 0);
-    if (ret != ACL_ERROR_NONE) {
-        LOG(INFO) << "Malloc host memory for numa:" << prop.location.id
-                  << " failed, try common allocate instead.";
-        prop.location.type = ACL_MEM_LOCATION_TYPE_HOST;
-        prop.location.id = 0;
-        ret = aclrtMallocPhysical(&handle, total_size, &prop, 0);
-        if (ret != ACL_ERROR_NONE) {
-            LOG(ERROR) << "Failed to allocate memory: " << ret;
-            return -1;
-        }
-    }
-    return 0;
-}
-#endif
-
 void *allocate_buffer_allocator_memory(size_t total_size,
                                        const std::string &protocol,
                                        size_t alignment) {
@@ -126,40 +84,9 @@ void *allocate_buffer_allocator_memory(size_t total_size,
         LOG(ERROR) << "Total size must be at least " << alignment;
         return nullptr;
     }
-#ifdef ASCEND_SUPPORT_FABRIC_MEM
-    if ((protocol == "ascend" || protocol == "ubshmem") && total_size > 0 &&
-        globalConfig().ascend_use_fabric_mem) {
-        aclrtDrvMemHandle handle = nullptr;
-        if (allocate_physical_memory(total_size, handle) != 0) {
-            return nullptr;
-        }
-        void *va;
-        auto ret = aclrtReserveMemAddress(&va, total_size, 0, nullptr, 1);
-        if (ret != ACL_ERROR_NONE) {
-            LOG(ERROR) << "Failed to reserve memory: " << ret;
-            (void)aclrtFreePhysical(handle);
-            return nullptr;
-        }
-        ret = aclrtMapMem(va, total_size, 0, handle, 0);
-        if (ret != ACL_ERROR_NONE) {
-            LOG(ERROR) << "Failed to map memory: " << ret;
-            (void)aclrtReleaseMemAddress(va);
-            (void)aclrtFreePhysical(handle);
-            return nullptr;
-        }
-        return va;
-    }
-#endif
-
-#ifdef USE_ASCEND_DIRECT
-    if (protocol == "ascend" && total_size > 0) {
-        void *buffer = nullptr;
-        auto ret = aclrtMallocHost(&buffer, total_size);
-        if (ret != ACL_ERROR_NONE) {
-            LOG(ERROR) << "Failed to allocate memory: " << ret;
-            return nullptr;
-        }
-        return buffer;
+#if defined(USE_ASCEND_DIRECT) || defined(USE_UBSHMEM)
+    if (protocol == "ascend" || protocol == "ubshmem") {
+        return ascend_allocate_memory(total_size, protocol);
     }
 #endif
 
@@ -201,38 +128,9 @@ void free_buffer_mmap_memory(void *ptr, size_t total_size) {
 }
 
 void free_memory(const std::string &protocol, void *ptr) {
-#ifdef ASCEND_SUPPORT_FABRIC_MEM
-    if ((protocol == "ascend" || protocol == "ubshmem") &&
-        globalConfig().ascend_use_fabric_mem) {
-        aclrtDrvMemHandle handle;
-        auto ret = aclrtMemRetainAllocationHandle(ptr, &handle);
-        if (ret != ACL_ERROR_NONE) {
-            LOG(ERROR) << "Failed to retain allocation handle: " << ptr;
-            return;
-        }
-        ret = aclrtUnmapMem(ptr);
-        if (ret != ACL_ERROR_NONE) {
-            LOG(ERROR) << "Failed to unmap memory: " << ptr;
-            return;
-        }
-        ret = aclrtReleaseMemAddress(ptr);
-        if (ret != ACL_ERROR_NONE) {
-            LOG(ERROR) << "Failed to release mem address: " << ptr;
-            return;
-        }
-        ret = aclrtFreePhysical(handle);
-        if (ret != ACL_ERROR_NONE) {
-            LOG(ERROR) << "Failed to free physical handle: " << handle;
-            return;
-        }
-        return;
-    }
-#endif
-
-#ifdef USE_ASCEND_DIRECT
-    if (protocol == "ascend") {
-        aclrtFreeHost(ptr);
-        return;
+#if defined(USE_ASCEND_DIRECT) || defined(USE_UBSHMEM)
+    if (protocol == "ascend" || protocol == "ubshmem") {
+        return ascend_free_memory(protocol, ptr);
     }
 #endif
