+++ mooncake-transfer-engine/src/transport/ascend_transport/CMakeLists.txt
@@ -4,16 +4,32 @@ if (USE_ASCEND)
 elseif (USE_ASCEND_HETEROGENEOUS)
 	file(GLOB ASCEND_SOURCES "heterogeneous_rdma_transport/*.cpp")
 elseif (USE_UBSHMEM)
-    file(GLOB ASCEND_SOURCES "ubshmem_transport/*.cpp")
+    file(GLOB ASCEND_SOURCES "ubshmem_transport/*.cpp" "*.cpp")
 else ()
-    file(GLOB ASCEND_SOURCES "ascend_direct_transport/*.cpp")
+    file(GLOB ASCEND_SOURCES "ascend_direct_transport/*.cpp" "*.cpp")
 endif ()
 
-if (USE_ASCEND_DIRECT)
-    set(ADXL_HEADER_PATH "${ASCEND_INCLUDE_DIR}/adxl/adxl_engine.h")
-    if(EXISTS "${ADXL_HEADER_PATH}")
-        file(READ "${ADXL_HEADER_PATH}" ADXL_CONTENT)
-        if("${ADXL_CONTENT}" MATCHES "Status TransferAsync")
+if (USE_ASCEND_DIRECT OR USE_UBSHMEM)
+    set(ACL_RUNTIME_HEADER_PATH "${ASCEND_INCLUDE_DIR}/acl/acl_rt.h")
+    if(EXISTS "${ACL_RUNTIME_HEADER_PATH}")
+        file(READ "${ACL_RUNTIME_HEADER_PATH}" ACL_RUNTIME_CONTENT)
+        if("${ACL_RUNTIME_CONTENT}" MATCHES "aclrtMemRetainAllocationHandle")
+            message(STATUS "Ascend support fabric mem.")
+            add_compile_definitions(ASCEND_SUPPORT_FABRIC_MEM)
+        else()
+            message(STATUS "Ascend do not support fabric mem.")
+        endif()
+    else()
+        message(WARNING "Acl runtime header file is not exist")
+    endif()
+    set(ADXL_ENGINE_HEADER_PATH "${ASCEND_INCLUDE_DIR}/adxl/adxl_engine.h")
+    if(EXISTS "${ADXL_ENGINE_HEADER_PATH}")
+        file(READ "${ADXL_ENGINE_HEADER_PATH}" ADXL_HEADER_CONTENT)
+        if("${ADXL_HEADER_CONTENT}" MATCHES "MallocMem")
+            message(STATUS "Ascend support adxl MallocMem.")
+            add_compile_definitions(ASCEND_SUPPORT_ADXL_MALLOC_MEM)
+        endif()
+        if("${ADXL_HEADER_CONTENT}" MATCHES "Status TransferAsync")
             message(STATUS "Check TransferAsync method exist.")
             add_compile_definitions(EXIST_ADXL_ASYNC_METHOD)
         else()
